requirements.txt
django<5
celery
djangorestframework
markdown
django-filter
scikit-learn
pydantic
requests